Restructure tests and imports
